requirements.txt
asgiref==3.7.2
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.2
django-jazzmin==3.0.0
django-rest-passwordreset==1.4.1
django-shortuuidfield==0.1.3
django-simple-history==3.7.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.6
incremental==24.7.2
Markdown==3.6
packaging==24.1
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
psycopg==3.2.1
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.2.1
pypng==0.20220715.0
python-decouple==3.8
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
service-identity==24.1.0
setuptools==71.1.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.3
zope.interface==7.0.3
